Adds keyboard shortcuts to any field that has a calendar:
   't'         input today's date
   '+' or '='  increment the date in the field by one day
   '-'         decrement the date in the field by one day

When the calendar is visible, the shortcuts play nicely with it. If the calendar is not visible they still work properly, which makes them useful for keyboard-only next action input. Pressing '+' when no date is entered in the field will set the date to tomorrow, and likewise '-' with no date entered will set the date to yesterday.

Closes #264



git-svn-id: http://www.rousette.org.uk/svn/tracks-repos/trunk@231 a4c988fc-2ded-0310-b66e-134b36920a42