Properly update the project page for hidden projects
Immediately after hiding a project, if a user marks a todo as done, they
would erroneously see a "There are no incomplete actions in this
project" banner under the open actions section.

This commit uses the active_or_hidden scope for the project todos in
order to fix this, since the todos are active but hidden due to the
project.

Fixes #1202