Removing some more intermediate values
Also removing the python-esque combination methods to set multiple
disparate variables from a single method by refactoring that to a
simpler method.

Reducing the complexity of computing running averages.

Note: It is not clear to me why the actions_done_lastyears_data
action is entirely unscoped on date. That seems to be the expected
behavior, so I left it alone.