fix calendar_test
todos in rest of month means all todos after next week in the current month. The test failed
if the todo was created in next week and that week being the last week in the month. In that
case the todo should not be returned by rest_of_month since it is in rest_of_week