Add a configuration setting for serving static assets
This new setting will allow us to control in `site.yml` whether or not
our static assets will be served by Rails or by the app server that is
proxying requests to Rails. This is set to `false` by default so that
the proxying server is responsible for serving the assets.