Simplify the return values when using basic auth
Returning a hash with explicit keys isn't strictly necessary with the
access methods we're using to get the values for those keys out of the
hash. Return an empty hash instead, simplifying the code.

Also remove the early return statements within the conditional. Those
are also unneeded since this is the last expression that's run in the
method.