Don't include a protocol or domain for relative_to_root
Tracks uses this javascript method in the application layout called
`relative_to_root` to get relative paths for a lot of Javascript
functions. However, it includes the fully qualified protocol and domain
for the tracks site in question. If you have Tracks proxied through SSL
with nginx, then you get a non-secure URL which causes the Javascript to
not work to fetch things like contexts or projects for autocomplete.

Instead of using the `root_url` helper that Rails provides that won't
know about SSL, just use a normal forward slash instead, so that you
still get a path that's relative to the root instead of a URL.