Add tests for the new object and fix a bug
Each of the individual query chunks has their own test, in addition to a
test for the full combination of parameters that could influence a
query.

There is also a bugfix for the tag query in here, since I want, as much
as possible, to have passing tests on every commit.